Better hi parsing
Previous code set some expectations on the highlight command output which not
always are true. It exptected `ctermbg` to be exactly 3 chars long and `guibg`
to be exactly 7 chars long. This is not true for many cases. Here's for example
my `SingColumn`:

    SignColumn     xxx term=standout ctermfg=11 ctermbg=8 guifg=Cyan guibg=Grey

New code sets a different sets of expectations that might be a bit better. New
code relies on the normalized output of `highlight` command. It parses output
and takes the part from "=" to the first space. This should cover the case of
varying length of numeric colours as well as named colours.